# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SmartBudget API 설계 명세",
    "version": "1.3.0",
    "description": "인증·서버 버전·가계부·소비 분석·OCR에서 구현해야 할 API 계약의 단일 기준인 OpenAPI 3.1 문서입니다. 2026-09-24 현재 Python 서버는 공통 code 응답, 서버 버전 조회와 계정 PATCH·DELETE를 포함한 인증 API를 구현했습니다. 월별 예산 분리, 거래·리포트·OCR API는 아직 구현되지 않았습니다. 구현이 완료될 때까지 실행 중 서버의 /docs와 /openapi.json은 현재 구현 상태를 보여주고, 이 파일은 목표 계약을 나타냅니다.\n\n날짜·일시 형식 규칙: 달력 날짜는 ISO 8601 YYYY-MM-DD, 정확한 시점은 한국 시간 오프셋을 포함한 YYYY-MM-DDTHH:mm:ss+09:00을 사용합니다. 거래 시각처럼 날짜와 분리된 선택 시각은 HH:mm:ss 또는 null을 사용합니다."
  },
  "paths": {
...
    "/api/v1/report/{id}": {
      "delete": {
        "tags": [
          "소비 분석"
        ],
        "summary": "분석 삭제",
        "operationId": "delete_api_v1_report_id",
        "description": "- 기존 소비 분석 리포트를 DB에서 제거하는 API입니다. 출력 포맷은 JSON입니다.\n- 사용자 ID는 첨부된 JWT 페이로드 정보로 확인합니다.\n\nJWT는 Authorization: Bearer <JWT> 헤더로 전달합니다. query나 본문의 token으로 대체하지 않습니다. 조회 조건은 query, 대상 ID는 경로로 전달하며 요청 본문은 사용하지 않습니다.\n\n요청자의 소유 항목만 대상으로 합니다. 지정한 ID의 항목이 없거나 다른 사용자 소유이면 동일한 오류 코드 404를 반환하며, 다른 사용자 항목의 존재 여부를 구분하여 알리지 않습니다.\n\n정의되지 않은 query 또는 본문 필드를 전달하거나 GET·DELETE 요청에 본문을 포함하면 오류 코드 400을 반환합니다. 숫자 필드는 문자열·불리언으로 대체할 수 없습니다.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "소비 분석 리포트의 식별 ID입니다. 1 이상의 정수입니다.",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "정상 처리됨.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "code",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "const": 200
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "null"
                    },
                    "code": {
                      "type": "string",
                      "pattern": "^[A-Z][A-Z0-9_]*$",
                      "description": "클라이언트가 분기 처리할 안정적인 응답 코드입니다."
                    }
                  },
                  "additionalProperties": false
                },
                "example": {
                  "status": 200,
                  "message": "The request has been accepted and processed.",
                  "data": null,
                  "code": "SUCCESS"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "개인정보 응답의 캐시 저장 금지"
              }
            }
          },
          "400": {
            "description": "필수 파라미터가 누락되었거나 요청 형식이 유효하지 않음. 경로 id가 1 이상의 정수가 아니면 오류 코드 400을 반환합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "code",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "const": 400
                    },
                    "message": {
                      "type": "string",
                      "description": "오류 내용을 설명하는 문자열입니다."
                    },
                    "data": {
                      "type": "null"
                    },
                    "code": {
                      "type": "string",
                      "pattern": "^[A-Z][A-Z0-9_]*$",
                      "description": "클라이언트가 분기 처리할 안정적인 응답 코드입니다."
                    }
                  },
                  "additionalProperties": false
                },
                "example": {
                  "status": 400,
                  "message": "The request is invalid.",
                  "data": null,
                  "code": "INVALID_REQUEST"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "개인정보 응답의 캐시 저장 금지"
              }
            }
          },
          "401": {
            "description": "토큰이 제공되지 않았거나 검증에 실패함.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "code",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "const": 401
                    },
                    "message": {
                      "type": "string",
                      "description": "오류 내용을 설명하는 문자열입니다."
                    },
                    "data": {
                      "type": "null"
                    },
                    "code": {
                      "type": "string",
                      "pattern": "^[A-Z][A-Z0-9_]*$",
                      "description": "클라이언트가 분기 처리할 안정적인 응답 코드입니다."
                    }
                  },
                  "additionalProperties": false
                },
                "example": {
                  "status": 401,
                  "message": "Authentication is required or the token is invalid.",
                  "data": null,
                  "code": "AUTHENTICATION_REQUIRED"
                }
              }
            },
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "type": "string",
                  "const": "Bearer"
                },
                "description": "Bearer 인증 안내"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "개인정보 응답의 캐시 저장 금지"
              }
            }
          },
          "404": {
            "description": "지정한 ID의 본인 소유 항목을 찾을 수 없음. 존재하지 않는 항목과 다른 사용자 소유 항목 모두 같은 오류 코드 404를 반환합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "code",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "const": 404
                    },
                    "message": {
                      "type": "string",
                      "description": "오류 내용을 설명하는 문자열입니다."
                    },
                    "data": {
                      "type": "null"
                    },
                    "code": {
                      "type": "string",
                      "pattern": "^[A-Z][A-Z0-9_]*$",
                      "description": "클라이언트가 분기 처리할 안정적인 응답 코드입니다."
                    }
                  },
                  "additionalProperties": false
                },
                "example": {
                  "status": 404,
                  "message": "The requested resource was not found.",
                  "data": null,
                  "code": "RESOURCE_NOT_FOUND"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "개인정보 응답의 캐시 저장 금지"
              }
            }
          },
          "503": {
            "description": "The service is temporarily unavailable. DB 잠금·일시 사용 불가. Uvicorn의 동시 요청 제한 응답까지 이 JSON 형식을 보장하지는 않습니다.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "개인정보 응답의 캐시 저장 금지"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "code",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "const": 503
                    },
                    "message": {
                      "type": "string",
                      "const": "The service is temporarily unavailable."
                    },
                    "data": {
                      "type": "null"
                    },
                    "code": {
                      "type": "string",
                      "pattern": "^[A-Z][A-Z0-9_]*$",
                      "description": "클라이언트가 분기 처리할 안정적인 응답 코드입니다."
                    }
                  },
                  "additionalProperties": false
                },
                "example": {
                  "status": 503,
                  "message": "The service is temporarily unavailable.",
                  "data": null,
                  "code": "SERVICE_UNAVAILABLE"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "get": {
        "tags": [
          "소비 분석"
        ],
        "summary": "분석 상세 조회",
        "operationId": "get_api_v1_report_id",
        "description": "Authorization: Bearer <JWT>로 인증한 사용자가 본인 소유 소비 분석 리포트 하나를 ID로 조회합니다. full_report를 포함한 전체 리포트를 반환합니다. 대상이 없거나 다른 사용자 소유이면 같은 오류 코드 404를 반환합니다. query와 요청 본문은 허용하지 않습니다.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "소비 분석 리포트의 식별 ID입니다. 1 이상의 정수입니다.",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "정상 처리됨.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "개인정보 응답의 캐시 저장 금지"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "code",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "const": 200
                    },
                    "code": {
                      "type": "string",
                      "pattern": "^[A-Z][A-Z0-9_]*$",
                      "description": "클라이언트가 분기 처리할 안정적인 응답 코드입니다."
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "minimum": 1,
                          "description": "소비 분석 리포트의 식별 ID입니다. 1 이상의 정수입니다."
                        },
                        "full_report": {
                          "type": "string",
                          "description": "LLM이 반환한 소비 분석 리포트의 전체 내용이며 Markdown 형식의 문자열입니다.",
                          "minLength": 1
                        },
                        "generated_at": {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^[0-9]{4}-[0-9]{2}-[0-9]{2}T(?:[01][0-9]|2[0-3]):[0-5][0-9]:[0-5][0-9]\\+09:00(?![\\s\\S])",
                          "description": "리포트의 최근 생성 또는 재생성이 완료된 한국 시간입니다. YYYY-MM-DDTHH:mm:ss+09:00 형식입니다.",
                          "example": "2026-09-16T14:30:00+09:00"
                        },
                        "period_type": {
                          "type": "string",
                          "enum": [
                            "daily",
                            "weekly",
                            "monthly",
                            "yearly"
                          ],
                          "description": "리포트의 기간 단위입니다."
                        },
                        "period_start": {
                          "type": "string",
                          "description": "리포트가 분석한 기간의 시작 날짜입니다. 한국 시간 기준 YYYY-MM-DD 형식이며 해당 날짜를 포함합니다.",
                          "format": "date",
                          "pattern": "^[0-9]{4}-[0-9]{2}-[0-9]{2}(?![\\s\\S])",
                          "example": "2026-09-01"
                        },
                        "period_end": {
                          "type": "string",
                          "description": "리포트가 분석한 기간의 종료 날짜입니다. 한국 시간 기준 YYYY-MM-DD 형식이며 해당 날짜를 포함합니다.",
                          "format": "date",
                          "pattern": "^[0-9]{4}-[0-9]{2}-[0-9]{2}(?![\\s\\S])",
                          "example": "2026-09-30"
                        },
                        "is_stale": {
                          "type": "boolean",
                          "description": "리포트 생성 후 분석 기간의 거래가 추가·수정·삭제되어 재생성이 필요한지 나타냅니다."
                        }
                      },
                      "required": [
                        "id",
                        "full_report",
                        "generated_at",
                        "period_type",
                        "period_start",
                        "period_end",
                        "is_stale"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "additionalProperties": false
                },
                "example": {
                  "status": 200,
                  "code": "SUCCESS",
                  "message": "The request has been accepted and processed.",
                  "data": {
                    "id": 193923,
                    "full_report": "# 월간 소비 분석 리포트\n\n분석 기간: 2026-09-01 ~ 2026-09-30\n\n## 요약\n\n- 총 지출: 40,500원\n- 식비: 40,500원\n\n기간 내 지출은 식비로 분류되었습니다.",
                    "generated_at": "2026-10-01T00:00:21+09:00",
                    "period_type": "monthly",
                    "period_start": "2026-09-01",
                    "period_end": "2026-09-30",
                    "is_stale": false
                  }
                }
              }
            }
          },
          "400": {
            "description": "필수 파라미터가 누락되었거나 요청 형식이 유효하지 않음. 경로 id가 1 이상의 정수가 아니면 오류 코드 400을 반환합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "code",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "const": 400
                    },
                    "message": {
                      "type": "string",
                      "description": "오류 내용을 설명하는 문자열입니다."
                    },
                    "data": {
                      "type": "null"
                    },
                    "code": {
                      "type": "string",
                      "pattern": "^[A-Z][A-Z0-9_]*$",
                      "description": "클라이언트가 분기 처리할 안정적인 응답 코드입니다."
                    }
                  },
                  "additionalProperties": false
                },
                "example": {
                  "status": 400,
                  "message": "The request is invalid.",
                  "data": null,
                  "code": "INVALID_REQUEST"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "개인정보 응답의 캐시 저장 금지"
              }
            }
          },
          "401": {
            "description": "토큰이 제공되지 않았거나 검증에 실패함.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "code",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "const": 401
                    },
                    "message": {
                      "type": "string",
                      "description": "오류 내용을 설명하는 문자열입니다."
                    },
                    "data": {
                      "type": "null"
                    },
                    "code": {
                      "type": "string",
                      "pattern": "^[A-Z][A-Z0-9_]*$",
                      "description": "클라이언트가 분기 처리할 안정적인 응답 코드입니다."
                    }
                  },
                  "additionalProperties": false
                },
                "example": {
                  "status": 401,
                  "message": "Authentication is required or the token is invalid.",
                  "data": null,
                  "code": "AUTHENTICATION_REQUIRED"
                }
              }
            },
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "type": "string",
                  "const": "Bearer"
                },
                "description": "Bearer 인증 안내"
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "개인정보 응답의 캐시 저장 금지"
              }
            }
          },
          "404": {
            "description": "지정한 ID의 본인 소유 항목을 찾을 수 없음. 존재하지 않는 항목과 다른 사용자 소유 항목 모두 같은 오류 코드 404를 반환합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "code",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "const": 404
                    },
                    "message": {
                      "type": "string",
                      "description": "오류 내용을 설명하는 문자열입니다."
                    },
                    "data": {
                      "type": "null"
                    },
                    "code": {
                      "type": "string",
                      "pattern": "^[A-Z][A-Z0-9_]*$",
                      "description": "클라이언트가 분기 처리할 안정적인 응답 코드입니다."
                    }
                  },
                  "additionalProperties": false
                },
                "example": {
                  "status": 404,
                  "message": "The requested resource was not found.",
                  "data": null,
                  "code": "RESOURCE_NOT_FOUND"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "개인정보 응답의 캐시 저장 금지"
              }
            }
          },
          "503": {
            "description": "The service is temporarily unavailable. DB 잠금·일시 사용 불가. Uvicorn의 동시 요청 제한 응답까지 이 JSON 형식을 보장하지는 않습니다.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                },
                "description": "개인정보 응답의 캐시 저장 금지"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "code",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "const": 503
                    },
                    "message": {
                      "type": "string",
                      "const": "The service is temporarily unavailable."
                    },
                    "data": {
                      "type": "null"
                    },
                    "code": {
                      "type": "string",
                      "pattern": "^[A-Z][A-Z0-9_]*$",
                      "description": "클라이언트가 분기 처리할 안정적인 응답 코드입니다."
                    }
                  },
                  "additionalProperties": false
                },
                "example": {
                  "status": 503,
                  "message": "The service is temporarily unavailable.",
                  "data": null,
                  "code": "SERVICE_UNAVAILABLE"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Authorization: Bearer <JWT> 헤더로 인증합니다."
      }
    },
...
  }
}